fix: make it build without breaking dev stuff
