+++ front/gatsby/package.json
@@ -17,17 +17,21 @@
     "@types/react-helmet": "^5.0.14",
     "@types/react-transition-group": "^4.2.4",
     "@types/shortid": "^0.0.29",
+    "@types/ws": "^7.2.1",
     "apollo-boost": "^0.4.7",
     "apollo-link-ws": "^1.0.19",
     "gatsby": "^2.19.21",
     "gatsby-plugin-manifest": "^2.2.42",
     "gatsby-plugin-react-helmet": "^3.1.18",
     "gatsby-plugin-typescript": "^2.1.23",
+    "graphql": "^14.6.0",
     "react": "^16.11.0",
     "react-dom": "^16.11.0",
     "react-helmet": "^5.2.1",
     "react-transition-group": "^4.3.0",
-    "shortid": "^2.2.15"
+    "shortid": "^2.2.15",
+    "subscriptions-transport-ws": "^0.9.16",
+    "ws": "^7.2.1"
   },
   "scripts": {
     "build": "gatsby build",
