--- front/gatsby/src/pages/404.tsx
@@ -1,14 +1,0 @@
-// Copyright 2018-2020 @paritytech/Nomidot authors & contributors
-// This software may be modified and distributed under the terms
-// of the Apache-2.0 license. See the LICENSE file for details.
-
-import React from 'react';
-
-export default function NotFoundPage(): React.ReactElement {
-  return (
-    <>
-      <h1>NOT FOUND</h1>
-      <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-    </>
-  );
-}
